Split behavior impl method metadata tests
